[nvim] Improve LSP tooling
Install Mason for LSP handling and setup LSPs for Python, Rust, and Lua.

Install tools for auto completion and reconfigure diagnostic options.

Modify keymaps for jumping to usages, and add keymappings for renaming
and jumping between diagnostics.

Thanks a lot to
https://rsdlt.github.io/posts/rust-nvim-ide-guide-walkthrough-development-debug/
which I essentially followed to setup all of this.