Get rid of bare git repo setup
Reorganise the structure of this repository to contain dotfiles in a
more manageable way. Write a simple helper script to symlink all the
dotfiles.

The only meaningful changes in this commit are the removal of the
`config` alias to access the bare repo and the move of the scripts
directory to `.local/scripts`, where it is now symlinked to.