[vdirsyncer] Setup vdirsyncer
This is for syncing my caldav calendar locally so I can use e.g. khal to
access it. In order to not directly add the urls in my config, I added a
simple script `cred` that reads out a python module containing arbitrary
credentials. I added the credentials module to .gitignore