[stagit] Add script to stack git branches
The script is still very basic. It parses the git log to get a stack of
branches until a specific source (`main` per default) and then offers
some commands to execute on each element of this stack. More complicated
things like pausing to fix rebase conflicts is not yet supported.